+++ lib/buffer/buffer_init.c
@@ -1,6 +1,6 @@
 #include "buffer.h"
 
-void buffer_init(buffer* b,int (*op)(),int fd,
+void buffer_init(buffer* b,ssize_t (*op)(),int fd,
 		 unsigned char* y,unsigned long int ylen) {
   b->op=op;
   b->fd=fd;
