+++ lib/buffer/buffer_mmapread_fd.c
@@ -1,7 +1,7 @@
 #include <buffer.h>
 #include <mmap.h>
 
-extern int buffer_dummyreadmmap();
+extern ssize_t buffer_dummyreadmmap();
 
 int buffer_mmapread_fd(buffer *b, int fd) 
 {
