+++ lib/buffer/buffer_prefetch.c
@@ -1,7 +1,7 @@
-#include <byte.h>
-#include <buffer.h>
+#include "byte.h"
+#include "buffer.h"
 
-extern int buffer_stubborn_read(int (*op)(),int fd,const char* buf, unsigned int len);
+extern ssize_t buffer_stubborn_read(ssize_t (*op)(),int fd,const char* buf, unsigned int len);
 
 int buffer_prefetch(buffer* b, unsigned long n) {
   if(b->p && b->p+n>=b->a)
